Send GPIO state to connected devices when set
Allows 2nd/3rd devices connected to the same ground station to see
GPIO changes that would otherwise not be visible to them.

This might cause small UI glitches if someone rapidly turns on/off
a bunch of GPIOs at the same time (race condition).

We might need to disable the other GPIO checkboxes briefly until the
state message arrives (with a small timeout to ensure it doesn't lock
up on interference if the message never arrives).