Add gstreamer camera stream type
This is the workhorse for a lot of different camera types that are
not worth writing custom code for yet, or at all.

In some cases like the Jetson this is the only way to use some camera
setups.

Note that this does not yet read the camera settings from the
settings file, because we do not use the older settings file, it
no longer makes any sense due to the way it was written. We no longer
have raspivid as the primary camera and then manually written
gstreamer pipelines as the 2nd camera. Instead we can have up to 4
different cameras of different kinds and they are all autodetected. In
addition the differences in the way they work are abstracted and will
have a standardized settings list rather than being tied to the way
the camera works like the older raspivid extraparams arguments were.

Because the settings support has not been merged yet, the camera
providers are all using 720p48 by default with 5mbps bitrate.

CBR is going to be enabled by default but the control for it is not
yet wired up, we may not even want to do that because VBR is unsafe.