Add autoconfigured camera support
This will load the system camera manifest, discover all of the
available cameras, and then start a new camera stream for each one
using whichever camera stream provider is necessary for that camera
type.

This uses the gstreamer provider for all of them at the moment because
the others aren't merged yet. Settings support is not merged either
due to the change in the way the settings system will work for
cameras from now on, we no longer can use settings tied to a particular
kind of camera like raspivid EXTRAPARAMS, nor can we assume that the
primary camera is a pi camera and the others are always USB, they
can be a mix of anything on any platform.